refactor: extract createFeedHash runtime logic from @scomp/types to @scomp/core (scomp-owz)
