--- packages/types/jest.config.cjs
@@ -1,5 +1,0 @@
-const { createJestConfig } = require("../../jest.base.cjs");
-
-module.exports = createJestConfig({
-  rootDir: __dirname,
-});
