+++ packages/types/package.json
@@ -7,6 +7,6 @@
   "scripts": {
     "build": "tsc -b tsconfig.json",
     "lint": "echo \"No lint configured for @scomp/types\"",
-    "test": "jest --config jest.config.cjs"
+    "test": "echo \"No tests configured for @scomp/types\""
   }
 }
