+++ packages/types/src/index.ts
@@ -8,16 +8,12 @@ export {
 } from "./network";
 
 export {
-  createRuntimeNeutralFeedHasher,
-  createFeedHash,
   SCOMP_CONTROL_PLANE_NAMESPACE,
-  type FeedHashOptions,
   type ScompPriorityClass,
   type ScompPriorityHint,
   type ScompPriorityIndex,
   type ScompTransportPriorityHints,
   type ScompClientInvokeOptions,
-  type ScompFeedHashFunction,
   type ScompControlPlaneDiscoverRequest,
   type ScompControlPlaneDiscoveredService,
   type ScompControlPlaneDiscoverResponse,
